requirements.txt
numpy==1.22.3
pandas==1.4.2
torch==1.11.0
torchvision
torchaudio
pyyaml
click
pytest
tqdm
# cudatoolkit